Move plugin to folder
