+++ application.fam
@@ -14,7 +14,7 @@ App(
     sources=[
       "*.c",
       "aeabi_uldivmod.sx",
-      "!plugin*.c",
+      "!plugin/*.c",
     ],
     fap_icon="icons/logo.png",
     fap_category="NFC",
@@ -46,5 +46,5 @@ App(
     apptype=FlipperAppType.PLUGIN,
     entry_point="plugin_wiegand_ep",
     requires=["seader"],
-    sources=["plugin_wiegand.c"],
+    sources=["plugin/wiegand.c"],
 )
