+++ seader_i.h
@@ -41,7 +41,7 @@
 #include <Payload.h>
 #include <FrameProtocol.h>
 
-#include "plugin_interface.h"
+#include "plugin/interface.h"
 #include <flipper_application/flipper_application.h>
 #include <flipper_application/plugins/plugin_manager.h>
 #include <loader/firmware_api/firmware_api.h>
